clean up headers (2/2)

+++ explorerwrapper/dbgprint.h
@@ -257,7 +257,7 @@ inline void* FindByString(uintptr_t baseaddress, const wchar_t* RefStr)
 	}
 	if (!StringAddress)
 	{
-		for (int i = 0; i < TextSize; i++)
+		for (size_t i = 0; i < TextSize; i++)
 		{
 			if (wcscmp((const wchar_t*)RefStr, (const wchar_t*)(TextSection + i)) == 0)
 			{
